Safari-611.1.10/LayoutTests/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'document.getElementById("myText1").id = "myText1"')
PASS
